Initial support for exploded views of deb822 entries
Provide an exploded_list view of the list where each entry represents
a single release file (1:n relation, vs n:m), matching legacy
sources.list entries.

On save, merge the exploded entries back together as far as
possible. This is currently lacking perhaps as we do not reorder
the entries to accommodate a change in the middle.

Change aptsources.distro to use the exploded view.

Change the remove() code to be able to remove exploded entries
from the SourcesList.

This makes software-properties recognize Ubuntu Sources correctly
when they are stored in multi-value deb822 entries, but modifying
them does not work so far, presumably due to the separation of
the privileged daemon from the graphical application.