distro: Handle .comps being a property, not updatable in-place
Modify a temporary copy and then assign it instead of trying
to just update the return value in place.

In an optimal world, we'd use some kind of proxy list object
here that would write modifications back to the underlying
object but that is also annoying hard to write correctly.