deb822: Improve parsing of section separators
Previously an empty file would have been considered to have one
empty section which was not helpful and if you separated paragraphs
by two lines instead of one you'd also get one paragraph for the
extra empty line.

This was _sort of_ useful in preserving variable number of lines
between sections, with this change paragraphs will always be separated
by one line when written back.

Or at least they should but it seems the code in File will currently
always insert two empty lines when doing __str__(). But then it seems
we don't use that __str__() in the actual writer so we are fine.